refactor(project structure): UI
