+++ apps/client/src/common/components/error-boundary/ErrorBoundary.jsx
